requirements.txt
Django~=5.2.1
psutil~=7.0.0
PyNaCl~=1.5.0
